feat: Day 1 complete
I probably didn't write that great of code let's be honest. Never used
Kotlin so this was basically a rough draft for how the rest of the year
will go. :)

Any who this was an interesting problem. I wonder if I spent more time
on it if there would be a more optimal approach that would have more dry
code but part 1 uses heaps because we always want the min. Part 2 uses a
map and list however I didn't spend enough time to think about
optimizing my iterations. It should still be O(n) but I wonder if n
could have been reduced (the constant that's dropped).